fastapi/requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.135.1
uvicorn==0.42.0
starlette==0.52.1

# ─── Validation & Settings ────────────────────────────────────────────────────
# Pinned to a stable release — never use pre-release (b1/b2/rc) in production.
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
annotated-types==0.7.0

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.43
asyncpg==0.30.0   # async PostgreSQL driver (used by SQLAlchemy async engine)
greenlet==3.1.1   # required by SQLAlchemy for async support
# psycopg2-binary is the *sync* driver — removed: all DB access is async via asyncpg

# ─── Auth ─────────────────────────────────────────────────────────────────────
python-keycloak==7.1.1

# ─── HTTP Client ──────────────────────────────────────────────────────────────
httpx==0.28.1

# ─── System Monitoring ────────────────────────────────────────────────────────
psutil==6.1.1

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.2.2
anyio==4.12.1
click==8.3.1
h11==0.16.0
idna==3.11
typing_extensions==4.15.0
